+++ functions/wordpress/postTypes/isHierarchicalPostType.js
