+++ functions/wordpress/postTypes/isValidPostType.js
