+++ lib/wordpress/_global/getPostTypeById.js
@@ -1,4 +1,4 @@
-import isHierarchicalPostType from '@/functions/wordpress/posts/isHierarchicalPostType'
+import isHierarchicalPostType from '@/functions/wordpress/postTypes/isHierarchicalPostType'
 import queryPageById from '@/lib/wordpress/pages/queryPageById'
 import queryPostById from '@/lib/wordpress/posts/queryPostById'
 import queryTeamById from '@/lib/wordpress/teams/queryTeamById'
