+++ lib/wordpress/_global/getPostTypeStaticPaths.js
@@ -1,5 +1,5 @@
-import isHierarchicalPostType from '@/functions/wordpress/posts/isHierarchicalPostType'
-import isValidPostType from '@/functions/wordpress/posts/isValidPostType'
+import isHierarchicalPostType from '@/functions/wordpress/postTypes/isHierarchicalPostType'
+import isValidPostType from '@/functions/wordpress/postTypes/isValidPostType'
 import {initializeWpApollo} from '@/lib/wordpress/connector'
 import {postTypes} from '@/lib/wordpress/_config/postTypes'
 import {gql} from '@apollo/client'
